dfs_traverse added, modularized and tested.

+++ c_src/map.h
@@ -2,9 +2,9 @@
 #ifndef MAP_H
 #define MAP_H
 
+#include <stdint.h>
 
-
-extern int map[15][19];
+extern int16_t map[15][19];
 
 extern int food_packets[5][2];
